+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Ability to define custom valid IPs to access webhook.
 
 ## [0.43.0] - 2017-04-17
 ### Added