+++ CHANGELOG.md
@@ -3,13 +3,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [Unreleased]
 ### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+### Security
+
+## [0.44.0] - 2017-05-05
+### Added
 - Ability to define custom valid IPs to access webhook.
 - Execute commands via cron, using `cron` action and `g` parameter.
 ### Changed
 - Remodelled the config array to a more flexible structure.
 - `bot_username` and `secret` are no longer vital parameters.
-### Deprecated
-### Removed
 ### Fixed
 - Initialise loggers before anything else, to allow logging of all errors.
 ### Security